kafka-python-audio-stream-producer/requirements/requirements.txt
kafka-python
sounddevice
SpeechRecognition
pyttsx3
pyaudio